fix #24: adicionado elemento de escrita no LCD Alterada forma de exibir caixa do elemento (Ao inves de chamar a funcao de desenhar caixa 3D e desativar o codigo que desenha a parte em 3D, chamada diretamente a funcao que exibe a caixa sem ser 3D) Adicionados objetos CAN no grupo de cores de comunicacao

+++ prefdialog.cpp
@@ -187,7 +187,7 @@ static void MakeControls(void)
 
     grouper = CreateWindowEx(0, WC_BUTTON, _("Portas Seriais"),
         WS_CHILD | BS_GROUPBOX | WS_VISIBLE,
-        GROUP_LEFT, (GROUP_HEIGHT * 5) + (PADDING_TOP * 6), GROUP_WIDTH, GROUP_HEIGHT * 1.5, PrefDialog, NULL, Instance, NULL);
+        GROUP_LEFT, (GROUP_HEIGHT * 5) + (PADDING_TOP * 6), GROUP_WIDTH, (int)(GROUP_HEIGHT * 1.5), PrefDialog, NULL, Instance, NULL);
     NiceFont(grouper);
 
 	// Projetos Recentes
